cls-zlog: define epoch for uninitialized objects
This patch creates the semantics that an unitialized object behaves as
though its current sealed epoch is -1. This means that operations with
epoch = 0 (for instance) will succeed. This is simply a way of avoiding
sealing epoch 0 and having all clients start at epoch 1 for no practical
reason when the log is created.

This means that calls to zlog_read against objects that don't exist
should return NOT_WRITTEN rather than -ENOENT. The zlog_read handler had
only the READ flag set on it, which resulted in the OSD returning
-ENOENT instead of the code defined by cls_zlog. We set WRITE on the
handler to enable our semantics, but this might be something rather
non-intuitive and could perhaps change in the OSD?

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>