mgr: monitor-controlled always on modules
the original version of the always-on mgr modules feature had an issue
such that if a user-enabled module was included in the always-on list of
modules then after an upgrade the module would remain in the list of
enabled modules (because it was still in the manager map). this results
in awkward ux since the modules cannot actually be disabled. this patch
addresses the issue by altering MgrMap to contain the always on modules,
and centralizes control of this list in the monitor.

the list of always on modules is versioned at compile time, and the
monitor will update the manager map when this verison is bumped. manager
daemons enable the union of user-enabled modules and the list of always
on modules. this is important so that a user-enabled module is not
disabled if it goes through a cycle of being in the always on list, and
then being removed through an upgrade sequence.

communicating the always on modules to the daemon enables the daemon to
provide the best error reporting when an always on module encounters an
issue (i.e. complain loudly).

Signed-off-by: Noah Watkins <nwatkins@redhat.com>